refactor: align context and hook module structure
