+++ src/components/Editor/ItemPropertiesPanel/ItemPropertiesPanel.tsx
@@ -4,7 +4,7 @@
  */
 
 import { useState } from 'react';
-import { useEditorContext } from '@/contexts/editor-context';
+import { useEditorContext } from '@/hooks/useEditorContext';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
