+++ src/contexts/EditorContextObject.ts
@@ -1,4 +1,4 @@
-import { createContext, useContext } from 'react';
+import { createContext } from 'react';
 import type { Dispatch } from 'react';
 import type { EditorAction, EditorState } from './EditorContext';
 
@@ -10,11 +10,3 @@ export interface EditorContextValue {
 export const EditorContext = createContext<EditorContextValue | undefined>(
   undefined,
 );
-
-export function useEditorContext() {
-  const context = useContext(EditorContext);
-  if (!context) {
-    throw new Error('useEditorContext must be used within EditorProvider');
-  }
-  return context;
-}
