+++ src/contexts/PostedTemplatesContextObject.ts
@@ -1,4 +1,4 @@
-import { createContext, useContext } from 'react';
+import { createContext } from 'react';
 import type { Dispatch } from 'react';
 import type {
   PostedTemplatesAction,
@@ -13,13 +13,3 @@ export interface PostedTemplatesContextValue {
 export const PostedTemplatesContext = createContext<
   PostedTemplatesContextValue | undefined
 >(undefined);
-
-export function usePostedTemplatesContext() {
-  const context = useContext(PostedTemplatesContext);
-  if (!context) {
-    throw new Error(
-      'usePostedTemplatesContext must be used within PostedTemplatesProvider',
-    );
-  }
-  return context;
-}
